Rejoined ensemble_manager_mod and data_structure_mod.  Data_structure_mod no longer exists.  data_structure_mod was made to avoid the circular dependency model_mod -> ensemble_manager_mod -> model_mod. I have removed this circular dependency by removing IO from ensemble_manager_mod and from assim_model_mod.  This IO (for regular old dart files - time + state vector) is now in state_vector_io_mod.
Put back in the option to write diagnostic files.  Spare copies are not written if you do this. Diagnostic file write is in its own module (the init_diag, etc. is still in assim_model_mod for now). There is a skeleton diagnostic routine that will probably be deprecated. Do we need a diagnostic write that can make a diagnostic file from state_structure_mod?

git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@8516 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 39341cf07e9ddc252f236a53defd2d01202149fb