correct the perturb copies routine to only perturb the ensemble memeber, move the check for missing_r8 inside a logical so it only gets executed if missing values are possible. remove some unused local variables.
still to be done:  1) figure out where this code actually belongs
and move it there.  2) there are 3 different namelists which have
a perturb amplitude.  3) the filter namelist now has both
'start_from_restart' and 'perturb_restarts'.  i'll agree that
'perturb_restarts' is much more helpful a name, but then we
should remove 'start_from_restart' because it implies doing
the same thing. 4) the call to the model perturb routine should
at least get the ensemble size.  5) we should give an example
of how to perturb variable by variable, or allow different
amplitudes by variable or something as a template for people
to follow if they do have specialized perturbation needs.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@8731 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 9224e4dfd391e53bcef0b86a2c66bd8a58e6cc7f