two changes:  first, the intel compiler was unhappy about specifying val_pressure(ens_size) as the function result value before ens_size was declared (it is an argument to this function) in get_val_pressure(). the "fix" was just to move it below the declaration line for ens_size.
second, fix another array out-of-bounds if you have
specified noise_sd > 0 in the namelist. the code looped
from 1 to size(array) but the actual index bounds go
from 0 to size-1 (or more if the halo size is > 1).
the fix is to use lbound() and ubound() to get the
lower and upper index values.  these loops will
include the halo region; if you want to operate
only on the interior of the array you'd have to 
get is,ie; js,je; ks,ke from the bgrid derived types.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@9196 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: ff7a07db886e2636d400eb42dbe1e9190831c15e