merging rma_state_structure_test_dir back into the rma_trunk.
Added test directory within io that contains a work directory where
you can build and compile test code. Test code files are located in
the test directory and build scripts are located in the work directory
along with a quick_build.csh which will compile all of the programs.
This is similar to the utilities test directory.

This contains a template for testing the state structure. Currently
this program just reads a namelist.  This program is a starting point
for unit testing for the state_structure_mod.

Also added in this branch is simple read write restart test program.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@9683 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 1e0772b9a422833ac012476473e6bde6e09701c2