Adding some of the more commonly used CF-Conventions to files created from scratch, and appending global attributes to file who have variables that are clamped.  The CF-Conventions that we will include are
 * units
 * long_name
 * short_name
 * _FillValue
 * missing_value
 * add_offset
 * scale_factor

If these variables exist we will store them in the state structure
and use this information when writting files.  

Note: add_offset and scale_factor are not supported at the moment,
and your program will crash.  This is because we need to unpack and
pack the data properly.  Currently no models use scale factor or 
offset.

Inflation and Standard Deviation Files do not include clamping
information.

One additional feature is having a global DART_file_information
string used to describe the file, and a DART_note for the mean
and standard deviation mentioning that the values were computed
prior to clamping. Additionally revision, rev. date and create date
are included in the global attributes.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10096 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a58344deb736812bb87e0cae8af555e369a41df6