Merging in CM1 branch into the rma_trunk.  Additional changes have been made to the path_names_* in order for quickbuild to sucessfully compile for bgrid_solo, cam, cm1, lorenz_96, mpas_atm, POP, ROMS and wrf.
Allocating 1d arrays instead of 2d arrays for forward,
and mean copies.  This is so Openmpi 1.10.0 will compile
and run.

Adding KIND_SURFACE_TEMPERATURE
       KIND_2M_TEMPERATURE
       KIND_10M_U_WIND_COMPONENT 
       KIND_10M_V_WIND_COMPONENT 
to the DEFAULT_obs_kind_mod.F90.  These are specific to CM1.

Changed the default preprocess namelist for wrf to match the
DART trunk. This is useful for bitwise testing when comparing
for bitwise tests.

Added the option to make cam filter with mpi and nompi.




git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10343 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d9ccd78499149f71d5b0b93a7d644262b31318bf