The observation converter for ROMS ocean obs to put them into DART obs_seq format.
This converter is different from others in that it assumes 
the input files contain the already-evaluated forward
operator values, output from an ensemble of ROMS models.  

the advantage is that the forward operators were computed 
at the time of the observation during the model run.

the disadvantage is that there is no interpolation code
in the model_mod for ROMS now, so we cannot compute the
posterior obs values.

this code is standalone, but using filter with these
obs requires the code in the obs_def module which 
handles obs_seq files that contain precomputed forward
operators.  (this will be committed soon but is not
committed at this time.)



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10464 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 1d9e280577b4e577c28a05eb6f5714ece21cedfa