clean up the comments around the fortran SELECTED_KIND sections, including some comments from the MPI standards on how to select a variable of the intended number of bytes.
replaced parmnamelength with varnamelength, and will be committing
other code that uses obstypelength only for obs strings and uses
varnamelength for things unrelated to observations.  the fortran 90/95
standard sets the max var name length as 31 (our 'preprocess' program
generates integer parameter variables from strings and the limit
applies there.)  apparently fortran 2003 raises that limit to 63 but
at this time we're sticking with 31.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10624 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: df87ccc47f93d46095b6572685ed24b70b7f5e76