Removed the use of a single file that had some metadata that was previously not present in the original test files. Turns out those test files are not representative and it is easy to have that metadata in the original rom_obs.nc which then passes it on to the verification observation files.
Because we can assume the verification obs files have all the required
metadata, we can simplify the logic a bit and I have simplified the
namelist.  I have also added the metadata to the input test data.

The change to the model_mod.f90 is to remove public USE statement that
is not needed and improved an error message about the assumptions of
the time string units. 



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10687 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 7dd1238e40faf7d5a72cc3d98490edcc1225dc6c