Turns out the string I was using to denote a successful model advance was not sufficient.  If the log file has a phrase like "Blowing-up: Saving latest model state into  RESTART file" the model did not complete successfully, even though the log file ends: "ROMS/TOMS: DONE"
I turned off the prior state-space inflation in the input.nml.template file.
This was causing our test case (with a not-very-realistic observation sequence file)
to blow up in the 4th cycle. Compounded by the script not 
detecting the failed model advance, DART was failing with a really misleading 
error message.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@10773 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: c9b734ea40108db39023109b971d1386878a3ede