add a namelist option to make a copy of the buffer before a broadcast (was already an option to do this for send/recv). this is expected to be needed only for mpi implementations which can't deal with array sections properly.
rearrange the code in send, recv, and broadcast to combine
sections in a more sane way depending on whether you are
making a copy or not.  support the copy option even if
you have to break large transfers up into multiple operations.

you should *not* need to make these copies; they cost time
and space.  however when trying to debug strange mpi problems
trying them out is a good initial test.

there is still an issue on what the max buffer limit should
be for a single operation; we currently use 2,000,000,000
(not 2 * 1024^3) for both send, recv, and broadcast.  
johnny says it might be necessary to drop this slighly for 
broadcast with some of the mpi implementations on cheyenne.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11952 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 9aacf635c20a567bbca2175d1c753f60d625ca15