change the handling for surface obs - interpolate from the z face array, not the z center array, which changes depending on the height of level 2.   also define surface obs as having a scale height of -log(1) instead of computing two values which may be almost but not quite exactly 1.
if the vertical location of an obs is at level 1, return
a vertical location of lower=1, upper=1 and fraction = 0.
this will work correctly with both 2d and 3d fields and
will not generate an out-of-bounds array access for 2d
fields (which have no level 2).  

support putting surface variables into the state vector
to compute forward operators for temperature, winds,
and specific humidity.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@11963 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 0d8f22b824e40e54bd8cd168e2f44f803d4bf50d