expanded the netcdf cover routines in netcdf_utilities_mod.f90. will track mapping between ncid and original filename and include filename in output error messages now.  does more checks on number of dims when getting variable sizes.  will check for the scale and offset attributes when getting and setting variable values and fail if they are present.  if you run into this please save me a copy of the file for an example??
removed the 'lanai_bitwise' option from assim_tools because
we are confident in the manhattan version and it's getting almost
impossible to reproduce the order of operations now between lanai
and manhattan.  also removed the lanai perturb routines from several
of the model_mods - it did the perturbation in the same order as
lanai (which called perturb on an ensemble-by-ensemble basis and
not as manhattan does with partial bits of the state for all ensemble 
members).  since we are ignoring lanai bitwise it simplifies the 
model_mod code to remove those uncalled routines.

made sure the assim_tools_mod.pf.f90 and the filter_mod.dopplerfold.f90
are in sync with the originals.

removed the model-specific model_mod_check.f90 source files for all
models except clm.  they were not being included by the path_names
files anymore but still existed in the models directory.  clm stays
but was cut down to just the routines which are unique to it.

fixed the perturb routines in wrf and mpas_atm.  they are hardly
ever used (a reasonable starting ensemble needs and some reasonable
internal consistency) but i use them sometimes just for testing.
the code wasn't handling multiple domains and the overall variable
mins/maxes correctly.

the rest of the model_mod changes were just to rename nc_redef to
nc_begin_define_mode, nc_enddef to nc_end_define_mode, and nc_sync
to nc_synchronize_file to be less inscrutable about what they were.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@12449 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 237c5dae3c48b5c8a8a8406e2ef413d3a2e0b9e2