add comments to the gps forward operator and converter files to remind us that the local gps forward operator doesn't need the 6 items of metadata that the non-local one does.  we could save memory and file space if we redefined the local and nonlocal gps obs to be different types.  (comments only, no code changes).
minor non-results-changing update to text_to_obs.f90 converter:
one of the compilers whines about a line that includes: a = b * -1
because there's something unnatural about a star and a minus next to
each other.  changed to:  a = b * (-1) and the minus sign is now
safely quarantined in parens.  

i also had an example input text file with a few more lines 
so i updated that as well.




git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@12566 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 85746e346707f92d5c8d9cf1a9dd593059c646a6