updates to the scripts used to convert gps radio occultation and gps ionospheric observations to dart obs_seq format.
renamed some existing scripts to be more clear about
what they do, and added some examples of how to run
long series of conversion days.  also added the scripts
i used to convert ionospheric obs (ionPrf data) to
obs_seq format.  these are also committed to the
openggcm branch so if/when we merge, this will already
exist on the rma_trunk.

the batch functions run on cheyenne and have
sections for PBS and LSF.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@12571 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: bd3b5a0737c5f015b54b47b1ef2e77f7f1c44ed7