get rid of compile errors.  this is an alternative version of the prepbufr converter that is not finished but is intended to work for prepbufr files which were converted (by dss) to netcdf format instead of bufr.
we stopped working on it when we found out that the
netcdf version didn't have all the info we needed
to create dart obs_seq files - but dss said they
will add the fields we need so this may be a good
converter to use someday.

in the meantime make sure the repo version can at
least build without error.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/rma_trunk@12909 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 77d4692ddca043aeec88d948fe419de7353c1bf7