improved readability, internal comments for suggestions.
This file uses all kinds of css formatting available from documentation/html/doc.css
This file was not originally part of the DART download, so the links to the presentations
are wrong. The URLs will change shortly in any event - as CISL transitions away from
hosting the www.image.ucar.edu  site.