quad utils cleanup; mpi smaller message send; new sort options
updated most of the comments in the quad utils code and made some
minor code cleanups.

if packing small arrays into a single buffer, include the actual
data size in the call to the broadcast.  this should improve the
mpi communication performance for small sends.

committing an improved test program for sort, and update the sort
code to fix a logic bug. in the expression:  if (a .and. b) then ...
the fortran standard allows a and b to be evaluated in any order.
if a is range checking for a valid index and b is using that index
e.g. if (a > 1 .and. b(a) > 0) then ...
it can cause an out-of-bounds array reference if a is < 1.
the sort module still has a public for the simple versions of
the sorts, just for timing tests.  they should not be used in
any actual code.  the insertion sort is faster in some cases
and is recommended if the array size favors it.