avoid compiler warnings; clean up use section on obs utility program
the null_mpi_utils version cannot implement send and receive, so they
always throw an error.  make the target array in receive be intent(inout)
instead of just intent(out) - this will avoid compiler warnings about
not setting an output value for this subroutine argument.  the entire
body of the subroutine is a call to the error handler which exits, so
this doesn't change anything about how the routine operates.

in the obs utilities dir, there is the start of a program to generate
a grid of obs.  it has multiple use lines for the same module, one
with 'only' and one without.  this generated unhappy messages from
the compiler.  it's fixed now.