Adding FORD-style description.
The description is the first block of !> comments before the program declaration.
An empty comment line terminates what FORD shows in the 'description' but the
rest of the comments appear in the full page for the unit. see advance_time for
a good example.

Also removed the useless version control block at the bottom of the files.

Replacing doxygen-style inline comments with FORD-style comments.

Generating the graphs takes about 8 minutes on my Mac.
Generating the search index also takes about 8 minutes.